refactor(editor compiler): Move comparison operator constants to a separate file
